+++ src/BT/ActionNode.hpp
@@ -142,7 +142,7 @@ protected:
     virtual State::StateType performAction() override;
     virtual State::StateType continueAction() override;
 
-    virtual void getLuaStatesHelper(std::vector<lua_State*>& v) override;
+    virtual void getLuaStatesHelper(std::unordered_set<lua_State*>& s) override;
 
 private:
     State::StateType performLuaScript(bool isContinuing);